+++ src/main/java/com/new_ton/domain/dto/productionpage/ProductTableDto.java
@@ -12,5 +12,5 @@ public class ProductTableDto {
     private String nameprod;
     private String sp;
     private Double percent;
-    private Integer mass;
+    private Double mass;
 }
